+++ modules/offline-validator/src/org/sbml/jsbml/validator/SBMLPackage.java
@@ -1,4 +1,5 @@
-package org.sbml.jsbml;
+package org.sbml.jsbml.validator;
+
 
 public enum SBMLPackage 
 {
@@ -71,5 +72,23 @@ public enum SBMLPackage
 	}
 	
 	return null;
+    }
+    
+    public static String convertIdToString(int errorId) {
+	SBMLPackage p = SBMLPackage.getPackageForError(errorId);
+	return p.toString() + "-" + String.format("%05d", errorId - p.offset);
+    }
+
+    public static int convertStringToId(String errorId) {
+	String[] blocks = errorId.split("-");
+
+	if (blocks.length == 1) {
+	    return Integer.parseInt(blocks[0]);
+	} else {
+	    SBMLPackage p = SBMLPackage.getPackageWithName(blocks[0]);
+	    int id = Integer.parseInt(blocks[1]);
+
+	    return id + p.offset;
+	}
     }
 }
