+++ modules/offline-validator/src/org/sbml/jsbml/validator/factory/FactoryManager.java
@@ -1,7 +1,8 @@
 package org.sbml.jsbml.validator.factory;
 
 import java.util.List;
-import org.sbml.jsbml.SBMLPackage;
+
+import org.sbml.jsbml.validator.SBMLPackage;
 
 public interface FactoryManager 
 {
