+++ modules/offline-validator/src/org/sbml/jsbml/validator/factory/L3V1FactoryManager.java
@@ -37,4 +37,6 @@ public class L3V1FactoryManager extends AbstractFactoryManager {
 	list.add(CORE_20617);
 	list.add(CORE_20623);
     }
+    
+    public void add
 }
